HUGE CHANGES to power states.
Changed The constants from before over to floats, since deltaTime is in seconds.

Added time based checks to the power change methods, the shield must be at full hp for the delay assigned by the constants for power requirements to drop.

Also added a time based check where a power must be low for longer than just a few milliseconds for inconsisent power suppliers like steam engines. (even if they supplied an overkill amount of power the shield didn't like them)
This could probably be abused with an ACB, swapping a motor on and off every like quater of a second, but IDC.
Stuff like that is why its an easy to change constant.